app/requirements.txt
asgiref==3.4.1
comtypes==1.1.10
cycler==0.10.0
Cython==0.29.24
Django==3.2.7
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
opencv-python==4.5.3.56
pandas==1.3.3
Pillow==8.3.2
pycocotools==2.0.2
pyparsing==2.4.7
pypiwin32==223
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2021.1
pywin32==301
six==1.16.0
sqlparse==0.4.2
torch==1.9.1+cu102
torchaudio==0.9.1
torchvision==0.10.1+cu102
typing-extensions==3.10.0.2
utils==1.0.1
